refactor: share circuit breaker defaults
